First changes from PR review
